+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+-   [`GenerateFormUrl()`](./docs/forms-client.md#GenerateFormUrl) method
 -   [`DecryptUserToken()`](./docs/forms-client.md#DecryptUserToken) method
 
 ## 1.12.0 (2020-11-12)