+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 -   [`GenerateFormUrl()`](./docs/forms-client.md#GenerateFormUrl) method
+-   [`EncryptUserToken()`](./docs/forms-client.md#EncryptUserToken) method
 -   [`DecryptUserToken()`](./docs/forms-client.md#DecryptUserToken) method
 
 ## 1.12.0 (2020-11-12)