+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 -   [`EncryptUserToken()`](./docs/forms-client.md#EncryptUserToken) method
 -   [`DecryptUserToken()`](./docs/forms-client.md#DecryptUserToken) method
 -   `crn2` and `crn3` to `FormSubmissionEventConfigration`
+-   `includeSubmissionIdInPdf` to `FormSubmissionEventConfigration`
 
 ## 1.12.0 (2020-11-12)